Fix PR notes, minor bug fixes
