+++ sdk/kiosk/src/index.ts
@@ -3,6 +3,5 @@
 
 export * from './tx/kiosk';
 export * from './tx/transfer-policy';
-export * from './tx/transfer-request';
 export * from './query/kiosk';
 export { ObjectArgument, objArg } from './utils';
